ATT (Audio to Transcription)/requirements_ffmpegfix.txt
openai-whisper==20231117
imageio-ffmpeg>=0.4.9
torch>=2.0.0; platform_system != 'Darwin' or platform_machine != 'arm64'
torch>=2.1.0; platform_system == 'Darwin' and platform_machine == 'arm64'
